+++ apps/core_admin/urls.py
@@ -10,13 +10,11 @@ urlpatterns = [
     path('atendimentos/desativacao/', views.desativar_atendimentos_ixc, name='desativar_atendimentos_ixc'),
     path('atendimentos/edicao-ixc/', views.editar_atendimentos_ixc, name='editar_atendimentos_ixc'),
     path('logins/edicao-ixc/', views.editar_logins_ixc, name='editar_logins_ixc'),
-    path('clientes/cadastro-ixc/', views.cadastrar_clientes_ixc, name='cadastrar_clientes_ixc'),
     path('restaurar-backup/', views.restore_backup, name='restore_backup'),
     path('teste-smtp/', views.smtp_test, name='smtp_test'),
     path('modelo/', views.download_template, name='download_template'),
     path('atendimentos/desativacao/modelo/', views.download_template_desativacao_atendimento, name='download_template_desativacao_atendimento'),
     path('atendimentos/edicao-ixc/modelo/', views.download_template_edicao_atendimento_ixc, name='download_template_edicao_atendimento_ixc'),
     path('logins/edicao-ixc/modelo/', views.download_template_edicao_login_ixc, name='download_template_edicao_login_ixc'),
-    path('clientes/cadastro-ixc/modelo/', views.download_template_cadastro_cliente_ixc, name='download_template_cadastro_cliente_ixc'),
     path('testar-api-ixc/', views.testar_api_ixc, name='testar_api_ixc'),
 ]